Move object detector to separate file.
