+++ src/vision/CMakeLists.txt
@@ -19,6 +19,11 @@ file(GLOB ${PROJECT_NAME}_SOURCES
 
 add_library(${PROJECT_NAME} STATIC ${${PROJECT_NAME}_SOURCES})
 
+
+add_dependencies(${PROJECT_NAME} 
+    device
+)
+
 target_link_libraries(${PROJECT_NAME} PUBLIC 
     ${OpenCV_LIBS}
     spdlog::spdlog
